Remove CountAvailableWorkers() logic (#31)
We tried to do a good thing here by adjusting the requested number of parallel
jobs depending on the actual number of workers that were available, but it would
appear that we are not actually getting the data back that we expect, so tend to
error out even when background workers are available.

Just rip this out instead of trying to be smarter than Postgres, and just let
Postgres error if it needs to when we try to do our thing.